Formatting size handled in util, switching files changes preview
